+++ dbase.py
@@ -1,49 +1,38 @@
+"""Create the axisorder database."""
+
 import os
 import sqlite3
 import sys
 from pathlib import Path
 
-from wfs20 import __path__
-from wfs20.io import GDAL_INSTALLED
-
-if GDAL_INSTALLED:
-    from osgeo import osr
+from pyproj.crs import CRS
 
 # 0,5,6 are hard to define.
-_OrienTable = {0: "xy", 1: "yx", 2: "yx", 3: "xy", 4: "xy", 5: "xy", 6: "xy"}
+_OrienTable_old = {0: "xy", 1: "yx", 2: "yx", 3: "xy", 4: "xy", 5: "xy", 6: "xy"}
+_OrienTable = {"west": "xy", "east": "xy", "north": "yx", "south": "yx"}
 
 
 def execute_query(conn, query):
+    """Execute a query."""
     cur = conn.cursor()
     try:
         cur.execute(query)
         conn.commit()
     except Exception as e:
-        print(f"The error '{e}' occurred")
+        sys.stdout.write(f"The error '{e}' occurred\n")
     cur.close()
 
 
-def execute_read_query(conn, query):
-    cur = conn.cursor()
-    r = None
-    try:
-        cur.execute(query)
-        r = cur.fetchall()
-        return r
-    except Exception as e:
-        print(f"The error '{e}' occurred")
-    cur.close()
-
-
-def _CreateAxisOrderDBASE():
+def create_database():
     """Create the axisorder database via the proj.db used by GDAL."""
     # Some locations
-    pyloc = os.path.dirname(sys.executable)
+    db_dir = os.environ.get("PROJ_DATA", os.environ.get("PROJ_LIB"))
+    if db_dir is None:
+        raise ConnectionError("Proj database was not found.")
+
     # database connections
-    proj = sqlite3.connect(
-        Path(pyloc, "Lib\\site-packages\\osgeo\\data\\proj\\proj.db")
-    )
-    conn = sqlite3.connect(Path(__path__[0], "data\\axisorder.db"))
+    proj = sqlite3.connect(Path(db_dir, "proj.db"))
+    conn = sqlite3.connect(Path(pwd, "src", "wfs20", "data", "axisorder.db"))
     # set cursor
     proj_cur = proj.cursor()
     create_table = """\
@@ -59,11 +48,11 @@ CREATE TABLE IF NOT EXISTS axisorder (
         for code in tuple(
             proj_cur.execute(f"SELECT code FROM {crs_type} WHERE auth_name = 'EPSG';")
         ):
-            srs = osr.SpatialReference()
-            srs.ImportFromEPSG(code[0])
+            srs = CRS.from_epsg(code[0])
             url = f"http://epsg.io/{code[0]}"
-            order = _OrienTable[srs.GetAxisOrientation(None, 0)]
-            print(srs.GetAxisOrientation(None, 0))
+            direction = srs.axis_info[0].direction
+            sys.stdout.write(f"Direction is: {direction}\n")
+            order = _OrienTable[direction]
             add_to_table = f"""\
 INSERT INTO
       axisorder ('auth','code','order','reference')
@@ -71,7 +60,7 @@ VALUES
       ("EPSG",{code[0]},'{order}','{url}')
 """
             execute_query(conn, add_to_table)
-            print(f"Succesfully added EPSG:{code[0]}")
+            sys.stdout.write(f"Succesfully added EPSG:{code[0]}\n")
             # clean up the srs
             srs = None
     # close all connections and cursors
@@ -81,4 +70,5 @@ VALUES
 
 
 if __name__ == "__main__":
-    _CreateAxisOrderDBASE()
+    pwd = Path(__file__).parent
+    create_database()
