+++ docs/.gitignore
@@ -1,4 +1,5 @@
 /.quarto/
+/_extensions/
 /_freeze/
 /_inv/
 /_site/
