--- docs/_extensions/machow/interlinks/.gitignore
@@ -1,3 +1,0 @@
-*.html
-*.pdf
-*_files/
