--- docs/_extensions/machow/interlinks/_extension.yml
@@ -1,7 +1,0 @@
-title: Interlinks
-author: Michael Chow
-version: 1.1.0
-quarto-required: ">=1.2.0"
-contributes:
-  filters:
-    - interlinks.lua
