--- docs/_extensions/mcanouil/iconify/_extension.yml
@@ -1,7 +1,0 @@
-title: Iconify support
-author: Mickaël Canouil
-version: 2.0.0
-quarto-required: ">=1.2.280"
-contributes:
-  shortcodes:
-    - iconify.lua
