--- docs/_extensions/quarto-ext/fontawesome/_extension.yml
@@ -1,7 +1,0 @@
-title: Font Awesome support
-author: Carlos Scheidegger
-version: 1.1.0
-quarto-required: ">=1.2.269"
-contributes:
-  shortcodes:
-    - fontawesome.lua
