--- docs/_extensions/quarto-ext/fontawesome/assets/css/latex-fontsize.css
@@ -1,30 +1,0 @@
-.fa-tiny {
-  font-size: 0.5em;
-}
-.fa-scriptsize {
-  font-size: 0.7em;
-}
-.fa-footnotesize {
-  font-size: 0.8em;
-}
-.fa-small {
-  font-size: 0.9em;
-}
-.fa-normalsize {
-  font-size: 1em;
-}
-.fa-large {
-  font-size: 1.2em;
-}
-.fa-Large {
-  font-size: 1.5em;
-}
-.fa-LARGE {
-  font-size: 1.75em;
-}
-.fa-huge {
-  font-size: 2em;
-}
-.fa-Huge {
-  font-size: 2.5em;
-}
