--- docs/_extensions/quarto-ext/include-code-files/_extension.yml
@@ -1,9 +1,0 @@
-title: Include Code Files
-author: Bruno Beaufils
-version: 1.0.0
-quarto-required: ">=1.2"
-contributes:
-  filters:
-    - include-code-files.lua
-
-
