+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools>=61.0.0", "wheel"]
+requires = ["setuptools>=61.0.0", "wheel", "pyproj"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools.packages.find]
@@ -28,7 +28,7 @@ description = "Small library to request geospatial data (WFS)"
 readme = "README.md"
 license = {text = "MIT"}
 
-requires-python = ">=3.9"
+requires-python = ">=3.11"
 dependencies = [
 "lxml",
 "requests",
@@ -62,36 +62,30 @@ classifiers = [
 # Specify the Python versions you support here. In particular, ensure
 # that you indicate whether you support Python 2, Python 3 or both.
 'Programming Language :: Python :: 3',
-'Programming Language :: Python :: 3.9',
-'Programming Language :: Python :: 3.10',
 'Programming Language :: Python :: 3.11',
+'Programming Language :: Python :: 3.12',
+'Programming Language :: Python :: 3.13',
 ]
 
 # Project's optional dependencies (development)
 [project.optional-dependencies]
-all = [
-    "setuptools>=61.0.0",
-]
 dev = [
     "pre-commit",
+    "pyproj",
     "ruff",
-    "wfs20[all, docs, io, test]",
+    "wfs20[docs, test]",
 ]
 docs = [
     "jupyter",
     "jupyter-cache",
+    "matplotlib",
     "quartodoc",
-    "wfs20[all, io]",
-]
-io = [
-    "gdal>=3.5",
 ]
 test = [
 	"pytest>=2.7.3",
 	"pytest-cov",
     "pytest-mock",
 	"responses",
-    "wfs20[all, io]",
 ]
 
 [project.urls]
@@ -104,7 +98,6 @@ wfs20 = "wfs20.cli:main"
 
 [tool.make_env]
 channels = ["conda-forge"]
-deps_not_in_conda = ["quartodoc"]
 
 ## Some stuff for pytest and coverage
 [tool.pytest.ini_options]
@@ -140,25 +133,100 @@ ignore_errors = true
 [tool.coverage.html]
 directory = ".cov"
 
-## Linting stuff
-[tool.black]
-line-length = 88
-target-version = ['py311']
-
 [tool.ruff]
 line-length = 88
 
+[tool.ruff.lint]
 # enable pydocstyle (E), pyflake (F) and isort (I), pytest-style (PT)
 select = ["E", "F", "I", "PT", "D"]
-ignore-init-module-imports = true
 ignore = ["B904", "D105", "D211", "D213", "D301", "E712", "E741"]
-exclude = ["docs"]
+exclude = ["docs/**"]
 
-[tool.ruff.per-file-ignores]
+[tool.ruff.lint.per-file-ignores]
 "test/**" = ["D103", "D100", "D104"]
 "test/conftest.py" = ["E402"]
 "src/wfs20/__init__.py" = ["E402", "F401", "F403"]
+"src/wfs20/driver/__init__.py" = ["F403"]
 "src/wfs20/cli.py" = ["D401"]
 
-[tool.ruff.pydocstyle]
+[tool.ruff.lint.pydocstyle]
 convention = "numpy"
+
+## Pixi related stuff
+[tool.pixi.project]
+channels = ["conda-forge"]
+platforms = ["linux-64", "win-64"]
+preview = ["pixi-build"]
+
+[tool.pixi.tasks]
+# Installation
+install = { depends-on = ["install-pre-commit"] }
+install-pre-commit = "pre-commit install"
+
+# Building
+build = { cmd = ["pyinstaller", "$PIXI_PROJECT_ROOT/.build/build.spec", "--distpath", "$PIXI_PROJECT_ROOT/bin", "--workpath", "$PIXI_PROJECT_ROOT/bin/intermediates"] }
+
+# Repo related tasks
+lint = { cmd = ["pre-commit", "run", "--all"] }
+
+# Docker related stuff (expand in future)
+docker-clean = { cmd = ["docker", "system", "prune", "-f"] }
+
+# Documentation generation
+docs = { depends-on = ["generate-api", "interlink-docs", "quarto-execute"] }
+docs-slim = { depends-on = ["interlink-docs", "quarto-render"] }
+docs-slim-execute = { depends-on = ["interlink-docs", "quarto-execute"] }
+generate-api = { cmd = ["python", "docs/_scripts/api.py", "docs/_quarto.yml"] }
+interlink-docs = { cmd = ["python", "docs/_scripts/inv.py", "interlinks", "docs/_quarto.yml"] }
+quarto-render = { cmd = ["quarto", "render", "docs"] }
+quarto-execute = { cmd = ["quarto", "render", "docs", "--execute"] }
+serve = { cmd = ["python", "-m", "http.server", "8000", "-d", "docs/_site"] }
+
+# Testing
+test = { cmd = ["pytest"] }
+test-lf = { cmd = ["pytest", "--lf", "--tb=short"] }
+test-cov = { cmd = ["pytest", "--verbose", "--cov=wfs20", "--cov-report", "xml"] }
+test-cov-local = { cmd = ["pytest", "--verbose", "--cov", "--cov-report", "html",] }
+
+# Clean up (some of) the temporary files
+clean = { depends-on = [
+  "clean-env-files",
+  "clean-bin",
+  "clean-build",
+  "clean-dist",
+  "clean-docs",
+] }
+clean-env-files = { cmd = ["rm", "-f", "environment.yml"] }
+clean-bin = { cmd = ["rm", "-rf", "bin"] }
+clean-build = {cmd = ["rm", "-rf", "build"] }
+clean-dist = {cmd = ["rm", "-rf", "dist"] }
+clean-docs = { cmd = ["rm", "-rf", "docs/_site"] }
+
+# Specify conda only deps
+[tool.pixi.dependencies]
+pip = "*"
+
+[tool.pixi.pypi-dependencies]
+wfs20 = { path = ".", editable = true }
+
+[tool.pixi.feature.docs.dependencies]
+quarto = "==1.5.57"
+
+# Extra python environments
+[tool.pixi.feature.py313.dependencies]
+python = "3.13.*"
+
+[tool.pixi.feature.py312.dependencies]
+python = "3.12.*"
+
+[tool.pixi.feature.py311.dependencies]
+python = "3.11.*"
+
+# Environments
+[tool.pixi.environments]
+default = { features = [] }
+dev = { features = ["dev", "docs", "test"] }
+docs = { features = ["docs",] }
+test-py313 = {features = ["py313", "test"]}
+test-py312 = {features = ["py312", "test"]}
+test-py311 = {features = ["py311", "test"]}
