+++ src/wfs20/__init__.py
@@ -1,4 +1,5 @@
 """WFS20."""
+
 ##################################################
 # Author: B.W. Dalmijn
 # E-mail: brencodeert@outlook.com
