+++ src/wfs20/driver/base.py
@@ -1,0 +1,1 @@
+"""Template for drivers."""
