+++ src/wfs20/driver/flatgeobuf.py
@@ -1,0 +1,1 @@
+"""FlatGeoBuf driver."""
