+++ src/wfs20/driver/geojson.py
@@ -1,0 +1,1 @@
+"""GeoJSON driver."""
