+++ src/wfs20/driver/geopackage.py
@@ -1,0 +1,1 @@
+"""GeoPackage driver."""
