+++ src/wfs20/wfs.py
@@ -1,11 +1,11 @@
 """Main submodule of WFS20."""
 
+from wfs20.build import build_service_meta
 from wfs20.crs import CRS
 from wfs20.error import WFSInternalError
-from wfs20.io import _WriteGeometries
+from wfs20.io import write_vector_ogr
 from wfs20.reader import DataReader, read_service
 from wfs20.request import create_get_request, service_url
-from wfs20.util import build_service_meta
 
 
 class WebFeatureService:
@@ -101,4 +101,4 @@ class WebFeatureService:
             raise WFSInternalError(
                 "Writing to file", "No features collected from WebFeatureService"
             )
-        _WriteGeometries(self.data_reader, driver, out)
+        write_vector_ogr(self.data_reader, driver, out)
